requirements.txt
fastapi~=0.89.1
uvicorn[standard]~=0.20.0
tensorflow~=2.11.0
numpy~=1.24.1
Pillow~=9.4.0
torch~=1.13.1